move fmod debugging functions to their own files
