+++ engine/audio/private/fmod_debug.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+void StartFMODDebugLogger();
