+++ engine/source/editor.cpp
@@ -128,12 +128,6 @@ void Editor::Draw(PerformanceTracker& performanceTracker, BloomSettings& bloomSe
 
     _entityEditor.renderSimpleCombo(_ecs->registry, _selectedEntity);
 
-    // if (ImGui::Begin("Entity Details"))
-    // {
-    //     DisplaySelectedEntityDetails();
-    // }
-    // ImGui::End();
-
     performanceTracker.Render();
     bloomSettings.Render();
